requirements.txt
pybaseball>=2.3.0
savant-extras>=0.3.0
pandas>=2.0
pyarrow>=14.0
google-cloud-bigquery>=3.0
google-cloud-bigquery-storage>=2.0
db-dtypes>=1.0
